requirements.txt
aerich==0.7.1
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
asgiref==3.5.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
black==23.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dictdiffer==0.9.0
Django==4.0.4
django-filter==21.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
frozenlist==1.3.1
idna==3.3
importlib-metadata==4.11.4
iso8601==1.1.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
Pillow==9.1.1
pkg_resources==0.0.0
platformdirs==3.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pyflakes==2.4.0
PyJWT==2.4.0
pypika-tortoise==0.1.6
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.10
yarl==1.8.1
zipp==3.8.0
